Files moved to game_engine folder and game_engine/gameplay renamed to game_engine/component
