+++ src/game_engine/activatable.h
@@ -1,5 +1,5 @@
-#ifndef SNAKE_GAMEENGINE_GAMEPLAY_ACTIVATABLE_H
-#define SNAKE_GAMEENGINE_GAMEPLAY_ACTIVATABLE_H
+#ifndef SNAKE_GAMEENGINE_ACTIVATABLE_H
+#define SNAKE_GAMEENGINE_ACTIVATABLE_H
 
 class activatable
 {
