+++ src/game_engine/component/behavior.cpp
