+++ src/game_engine/component/component_not_found.h
@@ -1,5 +1,5 @@
-#ifndef SNAKE_GAMEENGINE_GAMEPLAY_COMPONENTNOTFOUND_H
-#define SNAKE_GAMEENGINE_GAMEPLAY_COMPONENTNOTFOUND_H
+#ifndef SNAKE_GAMEENGINE_COMPONENT_COMPONENTNOTFOUND_H
+#define SNAKE_GAMEENGINE_COMPONENT_COMPONENTNOTFOUND_H
 
 #include <stdexcept>
 #include <string>
