+++ src/game_engine/component/updatable.h
@@ -1,0 +1,10 @@
+#ifndef SNAKE_GAMEENGINE_COMPONENT_UPDATABLE_H
+#define SNAKE_GAMEENGINE_COMPONENT_UPDATABLE_H
+
+class updatable
+{
+public:
+    virtual void update() = 0;
+};
+
+#endif
