--- src/game_engine/gameplay/updatable.h
@@ -1,10 +1,0 @@
-#ifndef SNAKE_GAMEENGINE_GAMEPLAY_UPDATABLE_H
-#define SNAKE_GAMEENGINE_GAMEPLAY_UPDATABLE_H
-
-class updatable
-{
-public:
-    virtual void update() = 0;
-};
-
-#endif
